ECDSA/requirements.txt
alabaster==0.7.12
astroid==2.4.2
Babel==2.8.0
bcrypt==3.2.0
breathe==4.20.0
capstone==4.0.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.3
cryptography==3.1
cycler==0.10.0
docutils==0.16
gmpy2==2.0.8
hmmlearn==0.2.3
idna==2.10
imagesize==1.2.0
intervaltree==3.1.0
isort==4.3.21
Jinja2==2.11.2
joblib==0.16.0
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
Mako==1.1.3
MarkupSafe==1.1.1
matplotlib==3.2.2
mccabe==0.6.1
numpy==1.19.0
packaging==20.4
paramiko==2.7.2
psutil==5.7.2
pycparser==2.20
pyelftools==0.26
Pygments==2.6.1
pylint==2.5.3
PyNaCl==1.4.0
pyparsing==2.4.7
pyserial==3.4
PySocks==1.7.1
python-dateutil==2.8.1
python-Levenshtein==0.12.0
pytz==2020.1
pyusb==1.0.2
requests==2.24.0
scikit-learn==0.23.2
scipy==1.5.0
six==1.15.0
snowballstemmer==2.0.0
sortedcontainers==2.2.2
Sphinx==3.2.1
sphinx-rtd-theme==0.5.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
threadpoolctl==2.1.0
toml==0.10.1
typed-ast==1.4.1
unicorn==1.0.2rc3
urllib3==1.25.10
wrapt==1.12.1
